+++ source/common/http/header_map_impl.cc
@@ -8,7 +8,6 @@
 #include "common/common/assert.h"
 #include "common/common/dump_state_utils.h"
 #include "common/common/empty_string.h"
-#include "common/common/utility.h"
 #include "common/singleton/const_singleton.h"
 
 #include "absl/strings/match.h"
@@ -274,20 +273,25 @@ void HeaderMapImpl::HeaderEntryImpl::value(const HeaderEntry& header) {
     return {&h.inline_headers_.name##_, &Headers::get().name};                                     \
   });
 
-/**
- * This is the static lookup table that is used to determine whether a header is one of the O(1)
- * headers. This uses a trie for lookup time at most equal to the size of the incoming string.
- */
-struct HeaderMapImpl::StaticLookupTable : public TrieLookupTable<EntryCb> {
-  StaticLookupTable() {
-    ALL_INLINE_HEADERS(INLINE_HEADER_STATIC_MAP_ENTRY)
+HeaderMapImpl::StaticLookupTable::StaticLookupTable() {
+  ALL_INLINE_HEADERS(INLINE_HEADER_STATIC_MAP_ENTRY)
+}
 
-    // Special case where we map a legacy host header to :authority.
-    add(Headers::get().HostLegacy.get().c_str(), [](HeaderMapImpl& h) -> StaticLookupResponse {
-      return {&h.inline_headers_.Host_, &Headers::get().Host};
-    });
-  }
-};
+const HeaderMapImpl::StaticLookupTable& HeaderMapImpl::staticLookupTable() const {
+  return ConstSingleton<StaticLookupTable>::get();
+}
+
+RequestHeaderMapImpl::RequestHeaderStaticLookupTable::RequestHeaderStaticLookupTable()
+    : StaticLookupTable() {
+  // Special case where we map a legacy host header to :authority.
+  add(Headers::get().HostLegacy.get().c_str(), [](HeaderMapImpl& h) -> StaticLookupResponse {
+    return {&h.inline_headers_.Host_, &Headers::get().Host};
+  });
+}
+
+const HeaderMapImpl::StaticLookupTable& RequestHeaderMapImpl::staticLookupTable() const {
+  return ConstSingleton<RequestHeaderStaticLookupTable>::get();
+}
 
 uint64_t HeaderMapImpl::appendToHeader(HeaderString& header, absl::string_view data,
                                        absl::string_view delimiter) {
@@ -305,17 +309,30 @@ uint64_t HeaderMapImpl::appendToHeader(HeaderString& header, absl::string_view d
 
 HeaderMapImpl::HeaderMapImpl() { inline_headers_.clear(); }
 
-HeaderMapImpl::HeaderMapImpl(
-    const std::initializer_list<std::pair<LowerCaseString, std::string>>& values)
-    : HeaderMapImpl() {
+HeaderMapImplPtr HeaderMapImpl::create(
+    const std::initializer_list<std::pair<LowerCaseString, std::string>>& values) {
+  auto new_header_map = std::make_unique<HeaderMapImpl>();
+  initFromInitList(*new_header_map, values);
+  return new_header_map;
+}
+
+HeaderMapImplPtr HeaderMapImpl::create(const HeaderMap& rhs) {
+  auto new_header_map = std::make_unique<HeaderMapImpl>();
+  copyFrom(*new_header_map, rhs);
+  return new_header_map;
+}
+
+void HeaderMapImpl::initFromInitList(
+    HeaderMapImpl& new_header_map,
+    const std::initializer_list<std::pair<LowerCaseString, std::string>>& values) {
   for (auto& value : values) {
     HeaderString key_string;
     key_string.setCopy(value.first.get().c_str(), value.first.get().size());
     HeaderString value_string;
     value_string.setCopy(value.second.c_str(), value.second.size());
-    addViaMove(std::move(key_string), std::move(value_string));
+    new_header_map.addViaMove(std::move(key_string), std::move(value_string));
   }
-  verifyByteSize();
+  new_header_map.verifyByteSize();
 }
 
 void HeaderMapImpl::updateSize(uint64_t from_size, uint64_t to_size) {
@@ -331,7 +348,7 @@ void HeaderMapImpl::subtractSize(uint64_t size) {
   cached_byte_size_ -= size;
 }
 
-void HeaderMapImpl::copyFrom(const HeaderMap& header_map) {
+void HeaderMapImpl::copyFrom(HeaderMapImpl& lhs, const HeaderMap& header_map) {
   header_map.iterate(
       [](const HeaderEntry& header, void* context) -> HeaderMap::Iterate {
         // TODO(mattklein123) PERF: Avoid copying here if not necessary.
@@ -344,17 +361,34 @@ void HeaderMapImpl::copyFrom(const HeaderMap& header_map) {
                                                          std::move(value_string));
         return HeaderMap::Iterate::Continue;
       },
-      this);
-  verifyByteSize();
+      &lhs);
+  lhs.verifyByteSize();
 }
 
-bool HeaderMapImpl::operator==(const HeaderMapImpl& rhs) const {
+namespace {
+
+HeaderMap::Iterate collectAllHeaders(const HeaderEntry& header, void* headers) {
+  static_cast<std::vector<std::pair<absl::string_view, absl::string_view>>*>(headers)->push_back(
+      std::make_pair(header.key().getStringView(), header.value().getStringView()));
+  return HeaderMap::Iterate::Continue;
+};
+
+} // namespace
+
+// This is currently only used in tests and is not optimized for performance.
+bool HeaderMapImpl::operator==(const HeaderMap& rhs) const {
   if (size() != rhs.size()) {
     return false;
   }
 
-  for (auto i = headers_.begin(), j = rhs.headers_.begin(); i != headers_.end(); ++i, ++j) {
-    if (i->key() != j->key().getStringView() || i->value() != j->value().getStringView()) {
+  std::vector<std::pair<absl::string_view, absl::string_view>> rhs_headers;
+  rhs_headers.reserve(rhs.size());
+  rhs.iterate(collectAllHeaders, &rhs_headers);
+
+  auto i = headers_.begin();
+  auto j = rhs_headers.begin();
+  for (; i != headers_.end(); ++i, ++j) {
+    if (i->key() != j->first || i->value() != j->second) {
       return false;
     }
   }
@@ -362,10 +396,10 @@ bool HeaderMapImpl::operator==(const HeaderMapImpl& rhs) const {
   return true;
 }
 
-bool HeaderMapImpl::operator!=(const HeaderMapImpl& rhs) const { return !operator==(rhs); }
+bool HeaderMapImpl::operator!=(const HeaderMap& rhs) const { return !operator==(rhs); }
 
 void HeaderMapImpl::insertByKey(HeaderString&& key, HeaderString&& value) {
-  EntryCb cb = ConstSingleton<StaticLookupTable>::get().find(key.getStringView());
+  EntryCb cb = staticLookupTable().find(key.getStringView());
   if (cb) {
     key.clear();
     StaticLookupResponse ref_lookup_response = cb(*this);
@@ -506,14 +540,14 @@ void HeaderMapImpl::setCopy(const LowerCaseString& key, absl::string_view value)
 
 uint64_t HeaderMapImpl::byteSize() const { return cached_byte_size_; }
 
-uint64_t HeaderMapImpl::byteSizeInternal() const {
+void HeaderMapImpl::verifyByteSizeInternalForTest() const {
   // Computes the total byte size by summing the byte size of the keys and values.
   uint64_t byte_size = 0;
   for (const HeaderEntryImpl& header : headers_) {
     byte_size += header.key().size();
     byte_size += header.value().size();
   }
-  return byte_size;
+  ASSERT(cached_byte_size_ == byte_size);
 }
 
 const HeaderEntry* HeaderMapImpl::get(const LowerCaseString& key) const {
@@ -554,7 +588,7 @@ void HeaderMapImpl::iterateReverse(ConstIterateCb cb, void* context) const {
 
 HeaderMap::Lookup HeaderMapImpl::lookup(const LowerCaseString& key,
                                         const HeaderEntry** entry) const {
-  EntryCb cb = ConstSingleton<StaticLookupTable>::get().find(key.get());
+  EntryCb cb = staticLookupTable().find(key.get());
   if (cb) {
     // The accessor callbacks for predefined inline headers take a HeaderMapImpl& as an argument;
     // even though we don't make any modifications, we need to cast_cast in order to use the
@@ -582,7 +616,7 @@ void HeaderMapImpl::clear() {
 }
 
 void HeaderMapImpl::remove(const LowerCaseString& key) {
-  EntryCb cb = ConstSingleton<StaticLookupTable>::get().find(key.get());
+  EntryCb cb = staticLookupTable().find(key.get());
   if (cb) {
     StaticLookupResponse ref_lookup_response = cb(*this);
     removeInline(ref_lookup_response.entry_);
@@ -605,7 +639,7 @@ void HeaderMapImpl::removePrefix(const LowerCaseString& prefix) {
     if (to_remove) {
       // If this header should be removed, make sure any references in the
       // static lookup table are cleared as well.
-      EntryCb cb = ConstSingleton<StaticLookupTable>::get().find(entry.key().getStringView());
+      EntryCb cb = staticLookupTable().find(entry.key().getStringView());
       if (cb) {
         StaticLookupResponse ref_lookup_response = cb(*this);
         if (ref_lookup_response.entry_) {
@@ -666,7 +700,7 @@ HeaderMapImpl::HeaderEntryImpl& HeaderMapImpl::maybeCreateInline(HeaderEntryImpl
 }
 
 HeaderMapImpl::HeaderEntryImpl* HeaderMapImpl::getExistingInline(absl::string_view key) {
-  EntryCb cb = ConstSingleton<StaticLookupTable>::get().find(key);
+  EntryCb cb = staticLookupTable().find(key);
   if (cb) {
     StaticLookupResponse ref_lookup_response = cb(*this);
     return *ref_lookup_response.entry_;